Misc batch changes
Note that this adds quite a bit of code that needs to be finished in future work.

Removed currently unused utility project AppDockEventSpy out of the solution - although the code is still kept.
Added initial work for using Window API hooks for OS windows open and close
Added test application
 - Windowed application that can be captured to exercise various features